Separate Translator into Translator and TranslationManager (#1709)
Take the code for generation of the translation mapping and put it in the SpatialVirtualWorkerTranslationManager.
